Update README.md
Github help guide says Github uses Markdown Text as documentation syntax and it's not meant to be used for final designed presentations such as websites. We can't do much formatting using Markdown as we could using HTML and CSS, so I couldn't indent some of the paragraphs and the reference list. Making the website or webpages is done using a different feature of GitHub called GitHub Pages. I'm going to learn it ASAIC